api/src/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-cors-headers==4.1.0
djoser==2.2.0
psycopg==3.1.9
drf-nested-routers==0.93.4
